+++ packages/fighting-components/_hooks/useFilterProps.ts
@@ -3,7 +3,10 @@ import type { ordinaryFunctionInterface as a } from '../_interface'
 
 /**
  * 过滤 props 需要使用的 props
- * 
+ *
+ * hooks 提出者：https://github.com/OnlyShadows
+ * hooks 开发者：https://github.com/Tyh2001
+ *
  * 需要传入两个泛型参数
  * @type { FPropsType } 传入的 props 类型
  * @type { needPropsType } 需要的 props 类型 / 返回值类型
@@ -11,29 +14,32 @@ import type { ordinaryFunctionInterface as a } from '../_interface'
  * @param need 需要的键集合
  * @returns 需要的 props
  */
-export const useFilterProps = <FPropsType, needPropsType>(props: FPropsType, need: string[]) => {
-  // 结果
-  const needProps: needPropsType = reactive({})
+export const useFilterProps = <FPropsType, needPropsType>(
+  props: FPropsType,
+  need: string[]
+) => {
+  type objectAny = Record<string, unknown>
 
-  /**
-   * 过滤 props
-   */
+  interface getPropsInterface {
+    (): needPropsType
+  }
+
+  const needProps: needPropsType | objectAny = reactive({} as const)
+
+  //  过滤 props
   const filterProps: a = (): void => {
     for (const key of need) {
-      if (Object.hasOwn(props, key)) {
-        needProps[key] = props[key]
+      if (Object.hasOwn(props as unknown as object, key)) {
+        needProps[key] = (props as unknown as objectAny)[key]
       }
     }
   }
 
-  /**
-   * 获取结果
-   * @returns 
-   */
-  const getProps = (): needPropsType => {
+  //  获取结果
+  const getProps: getPropsInterface = (): needPropsType => {
     filterProps()
-    return needProps
+    return needProps as needPropsType
   }
 
-  return { getProps }
+  return { getProps } as const
 }